Add files via upload
This is as far as I have developed this piece of software.
It contains certain functionalities like:
-Farming
-Invader Hunt
-Relocating around the in game map
-Unlocking tasks
-Yielding

The game uses image recognition to figure out where to click however due to different aspect ratios on different platforms it has uncertainties in some cases.
Game updates have also been released since the last update of this code, however due to the modular design, changes should be easily implemented. 
I have not included the "pics" folder as it probably goes against policies.